refactors models : parsers :: new package lock json
