+++ packages/kauri-wallet/src/app/services/api/api.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { ApiModel } from './api.model';
+import { ApiModel } from './models/api.model';
 import { environment } from '../../../environments/environment';
 import { DataService } from '../data/data.service';
-import { TxSendModel } from './tx-send.model';
+import { TxSendModel } from './models/tx-send.model';
 import { parseRawTransaction } from './parsers/raw-transaction-parser';
 
 @Injectable()
@@ -16,6 +16,7 @@ export class ApiService {
     getRawTransactions(apiModel: ApiModel) {
 
       const headers_object = new HttpHeaders();
+
       headers_object.append('Content-Type', 'application/json');
 
       const httpOptions = {
@@ -23,20 +24,20 @@ export class ApiService {
       };
 
       const txSendModel: TxSendModel = {} as TxSendModel;
+
       txSendModel.transactions = [];
       txSendModel.transactions.push(apiModel);
 
       const endpoint = environment.apiBase + 'getrawtransactions';
 
         return this.http.post(endpoint, txSendModel, httpOptions)
-          .subscribe(
-          (response) => {
+          .subscribe((response) => {
 
             const txModel = parseRawTransaction(response);
 
             this._dataService.rawTransactions = txModel;
-            },
-          (error) => console.log(error)
+
+            }, (error) => console.log(error)
         );
     }
