+++ packages/kauri-wallet/src/app/services/api/parsers/raw-transaction-parser.spec.ts
@@ -1,5 +1,5 @@
 import { ScriptPubKeyData, ScriptSigData, TxAddress, TxModel,
-  TxsData, TxVerbose, VinData, VOutData } from '../txModel';
+  TxsData, TxVerbose, VinData, VOutData } from '../models/txModel';
 import {
   parseAddressItem, parseRawTransaction, parseScriptPubKeyData,
   parseScriptSigData, parseTransactions, parseVerbose, parseVinData,
