+++ packages/kauri-wallet/src/app/wallet/wallet.component.html
@@ -4,7 +4,7 @@
 
   <h1>Wallet Component</h1>
 
-  <select aria-label="Currency" id="currency" name="currency" #currency>
+  <select aria-label="Select Currency" id="currency" name="currency" #currency>
     <option value="">---Select Currency---</option>
     <option value="NAV">NAV</option>
     <option value="BTC">BTC</option>
